patholens/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
opencv-python>=4.8.0
scikit-image>=0.21.0
openai>=1.0.0
python-dotenv>=1.0.0
scipy>=1.10.0
certifi>=2023.0.0
